title: Main Structure
type: implement / fix

the folder structure is defined. and index.vue file as the main route is here. 4 components are created. the main structure of the home page is also made; a simple input with 2 buttons (made by Button component) for adding and deleting the task.
the navbar consists of 2 buttons for controlling the theme and user on the right, and a hamburger menu button on the right. In next versions there will be a navbar on the left as a form of drawer navbar. none of the logic is implemented and it was all just styling and main structure.