title:
type: feat / fix / refactor

For storing the task's data and for offline usage, I decided to go with indexedDB. For better modularity and for the sake of reusable data, I made composable for each of the CRUD functionalities and the initialization of the database. This feature is still in development and not finished yet.