title: Storage solution
type: feat

Storing tasks using indexedDB is now fully functional. tasks get stored inside a database called "taskManager". the objectStore is also called "Tasks". The database initiates with the mounting of MainContainer component. Add button does add the task but still no getting tasks or displaying them (will be added in the next commit).
initDB composable, responsible for initialization of the database, is fixed as well. There were some issues with incompatibility of the state.taskData, which was caused by the reactivity of vue itself, but ultimately, with taking a copy from that object, it was resolved.