title: Fetching tasks from indexedDB and task card's styling
type: feat / refactor / styling

Now with adding each new task, all the tasks are fetched from the database and shown inside the DOM. Also when refreshing the page, the same happens.
All composable files for database interactions are refactored to .ts format; Although not completely implemented and converted, but initDB, addTask and getAllTasks files are refactored (could use more time).
The taskCard component is now styled and structured. Although it's not complete yet.