title: Refactoring indexedDB composables.
type: refactor / styling

The global interface for Task type was refactored; now all the global types that need to be available to many files, are stored inside global.d.ts file.
The useHead method used in app.vue file to rename the title, is now removed; Instead the title is set in the nuxt config file.
CustomBtn component had many unnecessary complications; therefore I converted it into a class-base component rather than a vue component. The styled for this button is now in the main.css file. All components that used this CustomBtn.vue component, were refactored to use the new class-base button components (Header.vue, MainContainer.vue, taskCard.vue etc).
The vueDraggable library for making the tasks draggable is installed. The implementation still needs more work. Also, the color mode plugin for Nuxt was installed as well. This plugin helps with the dark theme mechanism for the project. Although the dark theme is not entirely implemented for all elements and components, but the main wrapper component's background color ( MainContainer.vue component) is effected by the changing theme. The button for toggling theme is functional now.
A secondary button for choosing a date for the task is created; Although not functional yet.
Custom shadows in tailwind config (in nuxt config file) were changed as well.