title: Task update capabilities and draggable feature
type: feature

Now with checking the checkbox, the status of the task goes form "pending" to "done". Also a tick mark appears as well.
Draggable feature introduced in the last commit, is now fully working, with the desired animation. Granted, it might get changed till the end of project but the overall functionalities is final.
Minor refactoring to the updateTask.ts file also took place.