feat: Deleting task functionality, changing the order of tasks when dragging (not finished). style: Better styling for checked tasks.
Deleting tasks feature is now functional
Styles of checked off tasks are much better. The background darkens, along with the text. Schedule button goes away and appears when uncheck the task.
After dragging a task and reordering them, the reorder gets saved but due to database complications, the reorder doesn't fully save and doesn't appear on the DOM. If it continues to block the progress with this feature, I might have to change the database structure... maybe. As mentioned earlier, this feature needs further tweaking.