feat: Reordering feature refactor: Changing task's object structure and database structure
Now, if you reorder tasks to your liking, it will save it for the next time.
This was achieved by adding a new property to the task's object called "order". Now this property is the keyPath for the indexedDB object store. Also the autoIncrement is turned off due to some complications with this new structure. A "subTask" property was also added for sub-task capabilities in the future.