feat: Sub-task capabilities
Dragging and dropping a task on top of another will make a sub-task. For better understanding and visualization, I changed the "ghost" style (the style of task cards when they're being dragged and the place they're going to be dropped).
Still the sub tasks don't get saved and need further tweaking.