Refactor: MainContainer component more abstracted, change of state interface to State, new component for the part above tasks' list
The name of "sate" interface in global.d.ts file, was causing some confusion and problems in the file. So changed it to "State", avoiding confusion and remaining readable.
MainContainer.vue component was so cluttered and not organized. I decided to make the top part of it into a separate component called "TitleBar". And the whole draggable component is also more abstracted; TasksList.vue component is now a reusable option for all your draggable needs.
Duplication problem is back. I resorted to clearing and updating tasks in object store. Seems to be the only viable solution for now.